requirements.txt
py-lightpack==2.0.0
